Refactored logging middleware and added option for stdout logging
