+++ app.js
@@ -20,62 +20,15 @@ var express = require( 'express' ),
 	helmet = require( 'helmet' ),
 	flash = require( 'express-flash' ),
 	app = express(),
-	bunyan = require('bunyan'),
-	bunyanMiddleware = require('bunyan-middleware'),
-	SyslogStream = require('bunyan-syslog-unixdgram'),
 	http = require( 'http' ).Server( app );
 
 var Options = require( __js + '/options' )();
 app.use( Options.load );
 
-// Bunyan logging
-var bunyanConfig = {
-	name: 'Membership-System',
-	streams: []
-};
-
-if (config.log != undefined) {
-	bunyanConfig.streams.push({
-		type: "rotating-file",
-		path: config.log,
-		period: '1d', // rotates every day
-		count: 7 // keeps 7 days
-	})
-}
-
-if (config.syslog == true) {
-	var streamOptions = {
-		path: '/var/run/syslog'
-	}
-
-	var supported = true;
-	switch ( process.platform ) {
-		case 'darwin':
-			streamOptions.path = '/var/run/syslog';
-			break;
-		case 'linux':
-			streamOptions.path = '/dev/log';
-			break;
-		default:
-			console.error( "syslog output only supported on Linux and OS X" );
-			supported = false
-	}
-
-	if ( supported ) {
-		var stream = new SyslogStream( streamOptions );
-		bunyanConfig.streams.push( {
-			level: 'debug',
-			type: 'raw',
-			stream: stream
-		} );
-	}
-}
-
-var requestLogger = bunyan.createLogger( bunyanConfig );
-
-app.use( bunyanMiddleware( { logger: requestLogger } ) );
-
 var app_loader = require( __js + '/app-loader' );
+
+// Add logging capabilities
+app.use( require( __js + '/logging' ) );
 
 // Use helmet
 app.use( helmet() );
